Add docs and example var for plugin registration
To make the life of a new operator of a vorto repository easier there now
is a section on configuration of 3rd party generators via the PLUGINS
variable.

There is also a link to one of the profile config files containing an
example array of plugins hosted on AWS, so the operator may gain some
insight on how this could look like.

Add reference to docker docs to getting started

The getting started guide should also feature a guide for operators
containing a link to the README of the docker deployment.

Signed-off-by: Alexander Wellbrock <a.wellbrock@mailbox.org>