Update the docker docs and variables
Comment http proxy settings from docker env

It is a more sane value to not set these variables per default with
dummy values. Instead these should be commented out but present for
reference.

Update readme md formatting with linter defaults

Remove plugin apiv1 variables from example

The variables VORTO_URL and SERVICE_URL are not needed anymore in vorto
0.11 and newer, since plugins are now registered via the PLUGINS env
variable.

Add docs and example var for plugin registration (#1992)

To make the life of a new operator of a vorto repository easier there now
is a section on configuration of 3rd party generators via the PLUGINS
variable.

There is also a link to one of the profile config files containing an
example array of plugins hosted on AWS, so the operator may gain some
insight on how this could look like.

Add reference to docker docs to getting started

The getting started guide should also feature a guide for operators
containing a link to the README of the docker deployment.

Signed-off-by: Alexander Wellbrock <a.wellbrock@mailbox.org>