2462_followup
Coverage: added new test for search with fine-tuned specifications (now includes private namespaces the authenticated user is a collaborator of)
Functional fix: added explicit sorting of search results by name at controller level, since the controller talks to the repository directly in this case (hence, no service-level sorting). This is probably useless as it could be sorted in the front-end, but helps a bit with test stability.

Signed-off-by: Menahem Julien Raccah Lisei <menahemjulien.raccahlisei@bosch-si.com>